Pre-load variables to prevent uncessary IO reads
This commit adds a function to preload variables commonly used variables
throughout the limited area program. If a mesh instatiated with mode 'r', its
variables will be pre-loaded.

Variables are pre-loaded into the mesh class itself, thus a variable that is
pre-loaded can be accessed via: `self.latCell` or `mesh.latCell`.

To handle the loop over variables in MeshHandler.subset_fields(...), pre-loaded
variables are added to a new dictionary in _load_vars()
(MeshHandler.variables)which is used as a lookup.  If the variables are present
in that MeshHandler.variables, then use those, if they are not, then the
variable has not been pre-loaded, so read it in from disk.

For MeshHandler.create_graph_file(...), preloading variables would alter the
structure of subset_fields to create class variables of nCells, nEdges,
nEdgesOnCell, cellsOnCell and cellsOnEdge. If another commit/PR is done to
rework create_graph_file, it may be worth looking into this issue.