Refactor inPoint nearest_cell out of mark_boundary
This commit moves finding the nearest cell to the specified inPoint outside of
mark_boundary. There are two reasons for moving this operation outside of
mark_boundary.

First being that finding the nearest_point should be a separate operation to
marking the boundary. Having it in mark_boundary makes it difficult to find,
and it is a separate operation then mark boundary.

Second, it will save some computation times if mark_boundary is called
repeatably if mark_boundary is to be called multiple times in the future (ie
the equatorial channel).