Set default NETCDF format to be NETCDF3_64BIT_OFFSET
This commit makes the regional mesh be of the format NETCDF3_64BIT_OFFSET by
default.

It also contains a hook from create_region, to mesh.create_file, so a user can
change the file format of the region by editing only the create_region file. A
future commit could update the argument parser in create_region to allow the
file version to be specified in the command line.