Determine name of output region fname by the input file name
This commit adds logic that will select the output name of the region file. If
the the input mesh contains static and only static, the output file will be
named `name.static.nc` likewise for a fname with grid and only grid, it will be
named `name.grid.nc`.

If the file contains both static and grid or something completely different,
then it will be named `name.region.nc`.

Also remove nCells from the graph.info file.