Updated circle method radius to be meters
Previously, the radius in the circle method for specifying a region was
specified in Kilometers, which was different from the distance specified

Previously, the circle method radius was specified in kilometers. This was
different than the distances used for the semi-major and semi-minor axis in the
ellipse method, which were specified as meters.

This commit changes it to be meters to be align with the ellipse method.