Subset fields dimensioned by time
This commit adds changes to successfully subset fields that are dimensioned by
Time. Before, if a field was dimensioned by Time, an error would occur because
the region masks (i.e. glbBdyCellIDs, glbBdyVertexIDs or glbBdyEdgeIDs) would
be subsetting the Time dimension instead of the nCells, nVertices or nEdges
dimension.

This commit fixes the above by checking to see if the time dimension is in the
variable that is being copied and using the region masks to subset the second
nCells, nVertices or nEdges dimension.