Remove pre-read of variables to handle sets of large files
The code removed in this commit was in place to try to achieve efficiency by
pre-loading the memory of the file before loading it with netCDF4 Dataset. With
the ability to subset multiple meshes in one call, this would cause a bus error
if the meshes were significantly large, as each mesh would be loaded at once,
rather than when it was needed.

At the end of `LimitedArea.gen_region()`, after the call to
`Mesh.subset_fields()`, the created mesh region file AND the mesh that was used
to created are both closed. This ensures that only one mesh (and its regional
subset) are loaded at once.