When opening mesh for reading, don't pre-read entire file into memory buffer
In the MeshHandler class, opening a mesh for reading previously pre-read the
entire file into a memory buffer, and then provided that buffer to
netCDF4.Dataset with the 'memory' keyword argument. For very large meshes,
however, this can lead to memory allocation errors, and so this commit simply
creates a new Dataset using the filename of the input mesh.